Move 'newentry/search' package to '/search'
