+++ app/src/main/java/org/resolvetosavelives/red/AppDatabase.kt
@@ -4,14 +4,14 @@ import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
 import android.arch.persistence.room.TypeConverters
 import org.resolvetosavelives.red.bp.BloodPressureMeasurement
-import org.resolvetosavelives.red.newentry.search.Gender
-import org.resolvetosavelives.red.newentry.search.Patient
-import org.resolvetosavelives.red.newentry.search.PatientAddress
-import org.resolvetosavelives.red.newentry.search.PatientPhoneNumber
-import org.resolvetosavelives.red.newentry.search.PatientPhoneNumberType
-import org.resolvetosavelives.red.newentry.search.PatientStatus
-import org.resolvetosavelives.red.newentry.search.PatientWithAddressAndPhone
-import org.resolvetosavelives.red.newentry.search.SyncStatus
+import org.resolvetosavelives.red.search.Gender
+import org.resolvetosavelives.red.search.Patient
+import org.resolvetosavelives.red.search.PatientAddress
+import org.resolvetosavelives.red.search.PatientPhoneNumber
+import org.resolvetosavelives.red.search.PatientPhoneNumberType
+import org.resolvetosavelives.red.search.PatientStatus
+import org.resolvetosavelives.red.search.PatientWithAddressAndPhone
+import org.resolvetosavelives.red.search.SyncStatus
 import org.resolvetosavelives.red.util.InstantRoomTypeConverter
 import org.resolvetosavelives.red.util.LocalDateRoomTypeConverter
 import org.resolvetosavelives.red.util.UuidRoomTypeConverter
