+++ app/src/main/java/org/resolvetosavelives/red/bp/sync/BloodPressureSync.kt
@@ -7,7 +7,7 @@ import io.reactivex.Single
 import io.reactivex.rxkotlin.toCompletable
 import io.reactivex.schedulers.Schedulers
 import org.resolvetosavelives.red.bp.BloodPressureRepository
-import org.resolvetosavelives.red.newentry.search.SyncStatus
+import org.resolvetosavelives.red.search.SyncStatus
 import org.resolvetosavelives.red.sync.SyncConfig
 import org.resolvetosavelives.red.sync.patient.DataPushResponse
 import org.resolvetosavelives.red.util.Just
