+++ app/src/main/java/org/resolvetosavelives/red/newentry/address/PatientAddressEntryScreen.kt
@@ -14,8 +14,8 @@ import kotterknife.bindView
 import org.resolvetosavelives.red.R
 import org.resolvetosavelives.red.TheActivity
 import org.resolvetosavelives.red.newentry.phone.PatientPhoneEntryScreen
-import org.resolvetosavelives.red.newentry.search.OngoingPatientEntry
 import org.resolvetosavelives.red.router.screen.ScreenRouter
+import org.resolvetosavelives.red.search.OngoingPatientEntry
 import org.resolvetosavelives.red.widgets.ScreenCreated
 import org.resolvetosavelives.red.widgets.setTextAndCursor
 import org.resolvetosavelives.red.widgets.showKeyboard
