+++ app/src/main/java/org/resolvetosavelives/red/newentry/personal/PatientPersonalDetailsEntryEvents.kt
@@ -1,6 +1,6 @@
 package org.resolvetosavelives.red.newentry.personal
 
-import org.resolvetosavelives.red.newentry.search.Gender
+import org.resolvetosavelives.red.search.Gender
 import org.resolvetosavelives.red.widgets.UiEvent
 
 data class PatientFullNameTextChanged(val fullName: String) : UiEvent
