+++ app/src/main/java/org/resolvetosavelives/red/newentry/phone/PatientPhoneEntryScreen.kt
@@ -12,8 +12,8 @@ import io.reactivex.schedulers.Schedulers
 import kotterknife.bindView
 import org.resolvetosavelives.red.R
 import org.resolvetosavelives.red.TheActivity
-import org.resolvetosavelives.red.newentry.search.OngoingPatientEntry
 import org.resolvetosavelives.red.router.screen.ScreenRouter
+import org.resolvetosavelives.red.search.OngoingPatientEntry
 import org.resolvetosavelives.red.widgets.ScreenCreated
 import org.resolvetosavelives.red.widgets.setTextAndCursor
 import org.resolvetosavelives.red.widgets.showKeyboard
