+++ app/src/main/java/org/resolvetosavelives/red/search/OngoingPatientEntry.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.newentry.search
+package org.resolvetosavelives.red.search
 
 data class OngoingPatientEntry(
     val personalDetails: PersonalDetails? = null,
