+++ app/src/main/java/org/resolvetosavelives/red/search/PatientPhoneNumberType.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.newentry.search
+package org.resolvetosavelives.red.search
 
 import org.resolvetosavelives.red.util.RoomEnumTypeConverter
