+++ app/src/main/java/org/resolvetosavelives/red/search/PatientSearchByPhoneScreenController.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.newentry.search
+package org.resolvetosavelives.red.search
 
 import io.reactivex.Observable
 import io.reactivex.ObservableSource
