+++ app/src/main/java/org/resolvetosavelives/red/search/PatientStatus.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.newentry.search
+package org.resolvetosavelives.red.search
 
 import com.squareup.moshi.Json
 import org.resolvetosavelives.red.util.RoomEnumTypeConverter
