+++ app/src/test/java/org/resolvetosavelives/red/sync/PatientSyncTest.kt
@@ -13,13 +13,13 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.mockito.Mockito.reset
-import org.resolvetosavelives.red.newentry.search.PatientAddress
-import org.resolvetosavelives.red.newentry.search.PatientRepository
-import org.resolvetosavelives.red.newentry.search.PatientWithAddressAndPhone
-import org.resolvetosavelives.red.newentry.search.SyncStatus.DONE
-import org.resolvetosavelives.red.newentry.search.SyncStatus.INVALID
-import org.resolvetosavelives.red.newentry.search.SyncStatus.IN_FLIGHT
-import org.resolvetosavelives.red.newentry.search.SyncStatus.PENDING
+import org.resolvetosavelives.red.search.PatientAddress
+import org.resolvetosavelives.red.search.PatientRepository
+import org.resolvetosavelives.red.search.PatientWithAddressAndPhone
+import org.resolvetosavelives.red.search.SyncStatus.DONE
+import org.resolvetosavelives.red.search.SyncStatus.INVALID
+import org.resolvetosavelives.red.search.SyncStatus.IN_FLIGHT
+import org.resolvetosavelives.red.search.SyncStatus.PENDING
 import org.resolvetosavelives.red.sync.patient.DataPushResponse
 import org.resolvetosavelives.red.sync.patient.PatientPullResponse
 import org.resolvetosavelives.red.sync.patient.PatientSync
